+++ Views/Components/PreviewPanel.xaml.cs
@@ -1,0 +1,14 @@
+using System.Windows.Controls;
+
+namespace SwitchBlade.Views.Components
+{
+    public partial class PreviewPanel : System.Windows.Controls.UserControl
+    {
+        public PreviewPanel()
+        {
+            InitializeComponent();
+        }
+
+        public Canvas PreviewCanvas => ThumbnailCanvas;
+    }
+}