+++ Views/Components/SearchBar.xaml.cs
@@ -1,0 +1,17 @@
+using System.Windows.Controls;
+
+namespace SwitchBlade.Views.Components
+{
+    public partial class SearchBar : System.Windows.Controls.UserControl
+    {
+        public SearchBar()
+        {
+            InitializeComponent();
+        }
+
+        public void FocusInput()
+        {
+            InputBox.Focus();
+        }
+    }
+}